requirements.txt
Babel==2.12.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
passlib==1.7.4
python-decouple==3.8
python-dotenv==1.0.0
somepackage==1.2.3
speaklater==1.3
typing_extensions==4.7.1
Werkzeug==2.3.7
WTForms==3.0.1
